episode announcements, clean up
