--- src/episode_count
@@ -1,1 +1,0 @@
-7
